This commit will break Windows further until the widget system is hooked in; the most immediate error will be a link error regarding show_error. On the bright side, this commit updates the entirety of ENIGMA's library to follow a stricter standard. It is now expected that all library functions include the header; I have removed all redundant declarations and ensured that no functions are implemented without a prior prototype from a header.
With exception of HaRRiKiRi's sources and var4_lua, all ENIGMA warnings have been resolved. ALURE, which still needs updated, and the audio codecs, which are large and not our code, still warn to hell when compiled with -Wall -O3, and do not follow ENIGMA's new standard. I have no plans for dealing with them.

While this commit (or, to be more precise, this new standard which should have been better-enforced from the beginning) ensures that functions have consistent return types with their declarations, it does not remove all unnecessary return values. I have removed them from older, larger libraries such as GL's drawing functions, but some functions likely still return 0 unnecessarily, and others are equally likely to take double where they should take bool, etc.

It may behoove us to, if and when we fix return types, ensure that all collision functions return ::variant. This way, (bool)collision_line(...) behaves the same in ENIGMA as GM.

git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@914 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Universal_System/fileio.h
@@ -44,6 +44,6 @@ bool    file_bin_rewrite(int fileid);
 void    file_bin_close(int fileid);           
 size_t  file_bin_size(int fileid);            
 size_t  file_bin_position(int fileid);        
-void    file_bin_seek(int fileid,int pos);        
+void    file_bin_seek(int fileid,size_t pos);        
 void    file_bin_write_byte(int fileid,unsigned char byte); 
 int     file_bin_read_byte(int fileid);
